feat: integrate profile button and session management in header component
