+++ app/[locale]/dashboard/content.tsx
@@ -1,16 +1,7 @@
 "use client";
 
-import React, { useState } from "react";
-import {
-  CheckCircle,
-  User,
-  Bell,
-  LogOut,
-  Search,
-  Home,
-  Settings,
-  HelpCircle,
-} from "lucide-react";
+import React from "react";
+import { CheckCircle } from "lucide-react";
 import { Session } from "next-auth";
 
 type Props = {
@@ -18,194 +9,26 @@ type Props = {
 };
 
 export const DashboardContent = ({ session }: Props) => {
-  const [username, setUsername] = useState("John Doe");
-  const [isProfileMenuOpen, setIsProfileMenuOpen] = useState(false);
-  const [isNotificationsOpen, setIsNotificationsOpen] = useState(false);
-
-  console.log(session);
-
-  const handleLogout = () => {
-    console.log("Logging out...");
-  };
+  const user = session?.user;
 
   return (
-    <div className="min-h-screen bg-gray-50 flex flex-col">
-      {/* Header/Navigation Bar */}
-      <header className="bg-white shadow">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between h-16">
-            {/* Logo and Brand */}
-            <div className="flex">
-              <div className="flex-shrink-0 flex items-center">
-                <div className="h-8 w-8 bg-indigo-600 rounded-md flex items-center justify-center">
-                  <Home className="h-5 w-5 text-white" />
-                </div>
-                <span className="ml-2 text-xl font-semibold text-gray-900">
-                  Ever Directory
-                </span>
-              </div>
-
-              {/* Desktop Navigation */}
-              <div className="hidden md:ml-6 md:flex md:items-center md:space-x-4">
-                <a
-                  href="#"
-                  className="px-3 py-2 rounded-md text-sm font-medium text-gray-900 bg-gray-100"
-                >
-                  Dashboard
-                </a>
-                <a
-                  href="#"
-                  className="px-3 py-2 rounded-md text-sm font-medium text-gray-500 hover:text-gray-900 hover:bg-gray-100"
-                >
-                  Projects
-                </a>
-                <a
-                  href="#"
-                  className="px-3 py-2 rounded-md text-sm font-medium text-gray-500 hover:text-gray-900 hover:bg-gray-100"
-                >
-                  Files
-                </a>
-                <a
-                  href="#"
-                  className="px-3 py-2 rounded-md text-sm font-medium text-gray-500 hover:text-gray-900 hover:bg-gray-100"
-                >
-                  Team
-                </a>
-              </div>
-            </div>
-
-            {/* Search Bar */}
-            <div className="flex-1 flex items-center justify-center px-2 lg:ml-6 lg:justify-end">
-              <div className="max-w-lg w-full lg:max-w-xs">
-                <label htmlFor="search" className="sr-only">
-                  Search
-                </label>
-                <div className="relative">
-                  <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
-                    <Search className="h-5 w-5 text-gray-400" />
-                  </div>
-                  <input
-                    id="search"
-                    name="search"
-                    className="block w-full pl-10 pr-3 py-2 border border-gray-300 rounded-md leading-5 bg-white placeholder-gray-500 focus:outline-none focus:placeholder-gray-400 focus:ring-1 focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-                    placeholder="Search"
-                    type="search"
-                  />
-                </div>
-              </div>
-            </div>
-
-            {/* Right side icons */}
-            <div className="flex items-center">
-              {/* Notifications dropdown */}
-              <div className="relative ml-3">
-                <button
-                  type="button"
-                  className="p-1 rounded-full text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-                  onClick={() => setIsNotificationsOpen(!isNotificationsOpen)}
-                >
-                  <span className="sr-only">View notifications</span>
-                  <Bell className="h-6 w-6" />
-                </button>
-
-                {/* Notification dropdown */}
-                {isNotificationsOpen && (
-                  <div className="origin-top-right absolute right-0 mt-2 w-80 rounded-md shadow-lg py-1 bg-white ring-1 ring-black ring-opacity-5 focus:outline-none">
-                    <div className="px-4 py-2 border-b">
-                      <h3 className="text-sm font-medium text-gray-900">
-                        Notifications
-                      </h3>
-                    </div>
-                    <div className="px-4 py-2 text-sm text-gray-500 text-center">
-                      No new notifications
-                    </div>
-                  </div>
-                )}
-              </div>
-
-              {/* Profile dropdown */}
-              <div className="relative ml-3">
-                <div>
-                  <button
-                    type="button"
-                    className="flex text-sm rounded-full focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-                    id="user-menu"
-                    aria-expanded="false"
-                    aria-haspopup="true"
-                    onClick={() => setIsProfileMenuOpen(!isProfileMenuOpen)}
-                  >
-                    <span className="sr-only">Open user menu</span>
-                    <div className="h-8 w-8 rounded-full bg-indigo-600 flex items-center justify-center text-white">
-                      {username.charAt(0)}
-                    </div>
-                  </button>
-                </div>
-
-                {/* Profile dropdown menu */}
-                {isProfileMenuOpen && (
-                  <div
-                    className="origin-top-right absolute right-0 mt-2 w-48 rounded-md shadow-lg py-1 bg-white ring-1 ring-black ring-opacity-5 focus:outline-none"
-                    role="menu"
-                    aria-orientation="vertical"
-                    aria-labelledby="user-menu"
-                  >
-                    <a
-                      href="#"
-                      className="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-                      role="menuitem"
-                    >
-                      <User className="mr-3 h-4 w-4 text-gray-400" />
-                      Your Profile
-                    </a>
-                    <a
-                      href="#"
-                      className="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-                      role="menuitem"
-                    >
-                      <Settings className="mr-3 h-4 w-4 text-gray-400" />
-                      Settings
-                    </a>
-                    <a
-                      href="#"
-                      className="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-                      role="menuitem"
-                    >
-                      <HelpCircle className="mr-3 h-4 w-4 text-gray-400" />
-                      Help Center
-                    </a>
-                    <div className="border-t border-gray-100"></div>
-                    <button
-                      onClick={handleLogout}
-                      className="flex w-full items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-                      role="menuitem"
-                    >
-                      <LogOut className="mr-3 h-4 w-4 text-gray-400" />
-                      Sign out
-                    </button>
-                  </div>
-                )}
-              </div>
-            </div>
-          </div>
-        </div>
-      </header>
-
+    <div className="min-h-screen flex flex-col">
       {/* Main Content */}
       <main className="flex-grow">
         <div className="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">
           {/* Welcome Card */}
-          <div className="px-4 py-6 sm:px-0">
-            <div className="bg-white overflow-hidden shadow rounded-lg">
+          <div className="px-4 py-6 sm:px-0 border rounded-lg border-gray-200">
+            <div className="overflow-hidden shadow rounded-lg">
               <div className="px-4 py-5 sm:p-6">
                 <div className="flex items-center">
-                  <div className="flex-shrink-0 bg-green-100 rounded-md p-3">
+                  <div className="flex-shrink-0 rounded-md p-3">
                     <CheckCircle className="h-6 w-6 text-green-600" />
                   </div>
                   <div className="ml-5">
-                    <h3 className="text-lg leading-6 font-medium text-gray-900">
-                      Welcome back, {username}
+                    <h3 className="text-lg leading-6 font-medium text-gray-900 dark:text-gray-100">
+                      Welcome back, {user?.name}
                     </h3>
-                    <p className="mt-1 text-sm text-gray-500">
+                    <p className="mt-1 text-sm text-gray-500 dark:text-gray-100">
                       You have successfully logged in to your account.
                     </p>
                   </div>
@@ -214,14 +37,13 @@ export const DashboardContent = ({ session }: Props) => {
             </div>
           </div>
 
-          {/* Content Placeholder - This is where you would add your future features */}
           <div className="px-4 py-6 sm:px-0">
-            <div className="border-4 border-dashed border-gray-200 rounded-lg h-96 flex items-center justify-center">
+            <div className="border-4 border-dashed border-gray-200 dark:border-gray-500 rounded-lg h-96 flex items-center justify-center">
               <div className="text-center">
-                <h3 className="text-lg font-medium text-gray-900">
+                <h3 className="text-lg font-medium text-gray-900 dark:text-gray-100">
                   Ready to get started
                 </h3>
-                <p className="mt-1 text-sm text-gray-500">
+                <p className="mt-1 text-sm text-gray-500 dark:text-gray-100">
                   This is where your application content will go.
                 </p>
                 <div className="mt-6">
@@ -237,27 +59,6 @@ export const DashboardContent = ({ session }: Props) => {
           </div>
         </div>
       </main>
-
-      {/* Footer */}
-      <footer className="bg-white">
-        <div className="max-w-7xl mx-auto py-4 px-4 sm:px-6 md:flex md:items-center md:justify-between lg:px-8">
-          <div className="flex justify-center space-x-6 md:order-2">
-            <a href="#" className="text-gray-400 hover:text-gray-500">
-              <span className="sr-only">Help</span>
-              <HelpCircle className="h-5 w-5" />
-            </a>
-            <a href="#" className="text-gray-400 hover:text-gray-500">
-              <span className="sr-only">Settings</span>
-              <Settings className="h-5 w-5" />
-            </a>
-          </div>
-          <div className="mt-8 md:mt-0 md:order-1">
-            <p className="text-center text-sm text-gray-400">
-              &copy; 2025 DirectoryApp. All rights reserved.
-            </p>
-          </div>
-        </div>
-      </footer>
     </div>
   );
 };
